Refractored Code
