+++ GAN/.ipynb_checkpoints/GAN-checkpoint.py
@@ -5,20 +5,10 @@ import torch.optim as optim
 from torchvision import datasets, transforms
 from torch.autograd import Variable
 from torchvision.utils import save_image
-# from sklearn.manifold import TSNE
 import matplotlib.pyplot as plt
 import numpy as np
-# from sklearn import svm
-# import sklearn.model_selection
-# import sklearn.metrics
-# from sklearn.metrics import accuracy_score
-# from sklearn.metrics import confusion_matrix
-# from sklearn.metrics import precision_recall_fscore_support
-# from sklearn.ensemble import VotingClassifier
 import math 
-# import tensorflow as tf
 from torch.autograd import Variable
-
 import os
 from os import listdir
 from os.path import isfile, join 
@@ -59,10 +49,9 @@ torch.backends.cudnn.benchmark = True
 from prettytable import PrettyTable
 import torch.nn.init as init
 from vgg import vgg19
-
-
-#custom Initializer
+from models import SRSN_RRDB, DiscriminativeNet, VGGFeatureExtractor
 from initializer import kaiming_normal_
+from utilities import plot_grad_flow, count_parameters, SobelGrad
 
 
 ####### Initialisation 
@@ -142,94 +131,6 @@ def weight_init(m):
             else:
                 init.normal_(param.data)
    
-
-
-### Creating function for Gradient Visualisation 
-def plot_grad_flow(result_directory,named_parameters,model_name): 
-    
-    ave_grads = []
-    layers = []
-    for n, p in named_parameters:
-        if(p.requires_grad) and ("bias" not in n):
-            layers.append(n)
-            ave_grads.append(p.grad.abs().mean())
-    plt.figure(figsize=(12,12))
-    plt.plot(ave_grads, alpha=0.3, color="b")
-    plt.hlines(0, 0, len(ave_grads)+1, linewidth=1, color="k" )
-    plt.xticks(range(0,len(ave_grads), 1), layers, rotation="vertical")
-    plt.xlim(xmin=0, xmax=len(ave_grads))
-    plt.xlabel("Layers")
-    plt.ylabel("average gradient")
-    plt.title("Gradient flow")
-    plt.grid(True)
-    plt.savefig(os.path.join(result_directory, model_name + "gradient_flow.png" ))
-    
-### Get all the children layers 
-def get_children(model: torch.nn.Module):
-    # get children form model!
-    children = list(model.children())
-    flatt_children = []
-    if children == []:
-        # if model has no children; model is last child! :O
-        return model
-    else:
-       # look for children from children... to the last child!
-       for child in children:
-            try:
-                flatt_children.extend(get_children(child))
-            except TypeError:
-                flatt_children.append(get_children(child))
-    return flatt_children
-    
-
-### Layer Activation in CNNs 
-
-    
-def visualise_layer_activation(model,local_batch,result_directory):
-    activation = {}
-    def get_activation(name):
-        def hook(model, input, output):
-            activation[name] = output.detach()
-        return hook
-    layer_name = 'conv1'
-    model.module.conv1.register_forward_hook(get_activation(layer_name))
-    output = model(local_batch)
-    act = activation[layer_name].squeeze()
-    print(act.shape)
-    #plot subplots for different images
-    for i in range(act[0].shape[0]):
-        output = im.fromarray(np.uint8(np.moveaxis(act[0][i].cpu().detach().numpy(), 0, -1))).convert('RGB')
-        output.save(os.path.join(result_directory,str(i)+'.png'))
-        #
-
-### Visualising Conv Filters
-def visualise_conv_filters(model,result_directory):
-    kernels = model.conv1.weight.detach()
-    print(kernels.shape)
-    # fig, axarr = plt.subplots(kernels.size(0))
-    # for i in range(kernels.shape[0]):
-    #     plt.savefig()
-    # for idx in range(kernels.size(0)):
-    #     axarr[idx].imsave(kernels[idx].squeeze(),result_directory + "1.png")
-    #
-    
-def count_parameters(model):
-    table = PrettyTable(["Modules", "Parameters"])
-    total_params = 0
-    for name, parameter in model.named_parameters():
-        if not parameter.requires_grad: continue
-        param = parameter.numel()
-        table.add_row([name, param])
-        total_params+=param
-    print(table)
-    print(f"Total Trainable Params: {total_params}")
-    return total_params
-
-#########################################################################################################################################################################################################
-
-
-
-
 ### Data Preparation ############################################
 ################################################################
 
@@ -400,290 +301,6 @@ def process_and_train_load_data():
 #     calculate_mean_std_dataset(testloader_urban)
     
     return trainloader, testloader_flickr,testloader_div,testloader_urban
-
-class Normalize:
-    """Applies the :class:`~torchvision.transforms.Normalize` transform to a batch of images.
-    .. note::
-        This transform acts out of place by default, i.e., it does not mutate the input tensor.
-    Args:
-        mean (sequence): Sequence of means for each channel.
-        std (sequence): Sequence of standard deviations for each channel.
-        inplace(bool,optional): Bool to make this operation in-place.
-        dtype (torch.dtype,optional): The data type of tensors to which the transform will be applied.
-        device (torch.device,optional): The device of tensors to which the transform will be applied.
-    """
-
-    def __init__(self, mean, std, inplace=False, dtype=torch.float, device='cpu'):
-        self.mean = torch.as_tensor(mean, dtype=dtype, device=device)[None, :, None, None]
-        self.std = torch.as_tensor(std, dtype=dtype, device=device)[None, :, None, None]
-        self.inplace = inplace
-        
-    def __call__(self, tensor):
-        """
-        Args:
-            tensor (Tensor): Tensor of size (N, C, H, W) to be normalized.
-        Returns:
-            Tensor: Normalized Tensor.
-        """
-        if not self.inplace:
-            tensor = tensor.clone()
-
-        tensor.sub_(self.mean).div_(self.std)
-        return tensor
-    
-class ToTensor:
-    """Applies the :class:`~torchvision.transforms.ToTensor` transform to a batch of images.
-    """
-
-    def __init__(self):
-        self.max = 255
-        
-    def __call__(self, tensor):
-        """
-        Args:
-            tensor (Tensor): Tensor of size (N, C, H, W) to be tensorized.
-        Returns:
-            Tensor: Tensorized Tensor.
-        """
-        return tensor.float().div_(self.max)
-################################################################
-
-class DiscriminativeNet(torch.nn.Module):
-    
-    def __init__(self):
-        super(DiscriminativeNet, self).__init__()
-        
-        self.conv1 = nn.Sequential(
-            nn.Conv2d(
-                in_channels=3, out_channels=32, kernel_size=6, 
-                stride=2, padding=2, bias=False
-            ),
-            nn.LeakyReLU(0.2, inplace=True)
-        )
-        self.conv2 = nn.Sequential(
-            nn.Conv2d(
-                in_channels=32, out_channels=64, kernel_size=6,
-                stride=2, padding=2, bias=False
-            ),
-            nn.LeakyReLU(0.2, inplace=True),
-            nn.BatchNorm2d(64)
-        )
-        self.conv3 = nn.Sequential(
-            nn.Conv2d(
-                in_channels=64, out_channels=128, kernel_size=4,
-                stride=2, padding=1, bias=False
-            ),
-            nn.LeakyReLU(0.2, inplace=True),
-            nn.BatchNorm2d(128)
-        )
-        self.conv4 = nn.Sequential(
-            nn.Conv2d(
-                in_channels=128, out_channels=256, kernel_size=4,
-                stride=4, padding=0, bias=False
-            ),
-            nn.LeakyReLU(0.2, inplace=True),
-              nn.BatchNorm2d(256)
-        )
-#         self.conv5 = nn.Sequential(
-#             nn.Conv2d(
-#                 in_channels=256, out_channels=512, kernel_size=4,
-#                 stride=2, padding=1, bias=False
-#             ),
-#             nn.LeakyReLU(0.2, inplace=True),
-#             nn.BatchNorm2d(512)
-#         )
-#         self.conv6 = nn.Sequential(
-#             nn.Conv2d(
-#                 in_channels=512, out_channels=256, kernel_size=4,
-#                 stride=2, padding=1, bias=False
-#             ),
-#             nn.BatchNorm2d(256)
-#         )
-        self.conv7 = nn.Sequential(
-            nn.Conv2d(
-                in_channels=256, out_channels=128, kernel_size=4,
-                stride=4, padding=0, bias=False
-            ),
-            nn.BatchNorm2d(128)
-        )
-        self.Fc1 = nn.Sequential(
-            nn.Linear(128*4*4, 512),
-            nn.LeakyReLU(0.2, inplace=True)
-            # nn.Sigmoid(),
-        )
-        self.Fc2 = nn.Sequential(
-            nn.Linear(512,1),
-            nn.Sigmoid()
-        )
-
-    def forward(self, x):
-        # Convolutional layers
-        x = self.conv1(x)
-        x = self.conv2(x)
-        x = self.conv3(x)
-        x = self.conv4(x)
-#         x = self.conv5(x)
-#         x = F.relu(self.conv6(x))
-        x = F.relu(self.conv7(x))
-        
-        # Flatten and apply sigmoid
-#         print(x.shape)
-        x = x.view(-1, 128*4*4)
-#         print(x.shape)
-        x = self.Fc1(x)
-#         print(x.shape)
-        x = self.Fc2(x)
-#         print(x.shape)
-        # print(x)
-        return x
-    
-class VGGFeatureExtractor(nn.Module):
-    def __init__(self,
-                 low_feature_layer=22,
-                 high_feature_layer = 36,
-                 use_bn=False,
-                 device=torch.device('cuda')):
-        super(VGGFeatureExtractor, self).__init__()
-        if use_bn:
-            model = torchvision.models.vgg19_bn(pretrained=True)
-        else:
-            model = torchvision.models.vgg19(pretrained=True)
-        self.features_low = nn.Sequential(*list(model.features.children())[:(low_feature_layer + 1)])
-        self.features_high = nn.Sequential(*list(model.features.children())[:(high_feature_layer + 1)])
-        for param in self.parameters():
-            param.requires_grad = False
-
-    def forward(self, x):
-        return self.features_low(x), self.features_high(x)
-    
-class SRSN_RRDB(nn.Module):
-    def __init__(self, input_dim=3, dim=128, scale_factor=4,scale_ratio=0.2):
-        super(SRSN_RRDB, self).__init__()
-#         self.up = nn.ConvTranspose2d(3,3, 1, stride=1,output_size=(512,512))
-        self.conv1 = torch.nn.Conv2d(3, 128, 9, 1, 4)
-        self.conv2 = torch.nn.Conv2d(128, 64, 5, 1,2)
-        self.RDB1 = ResidualInResidualDenseBlock(64, 64, 0.2)
-        self.RDB2 =ResidualInResidualDenseBlock(64, 64, 0.2)
-        self.RDB3 = ResidualDenseBlock(64, 64, 0.2)
-        self.RDB4 = ResidualDenseBlock(64, 64, 0.2)
-        self.RDB5 = ResidualDenseBlock(64, 64, 0.2)
-        self.RDB6 = ResidualDenseBlock(64, 64, 0.2)
-        
-        self.up = torch.nn.Upsample(scale_factor=4, mode='bicubic')
-        self.conv3=torch.nn.Conv2d(64, 16, 3, 1, 1)
-        self.conv4=torch.nn.Conv2d(16, 3, 1, 1, 0)
-        self.conv5=torch.nn.Conv2d(64*6, 64, 1, 1, 0)
-        
-#         self.act1 = nn.PReLU()
-#         self.act2 = nn.PReLU()
-        
-#         self.bn = torch.nn.BatchNorm2d(3)
-        
-        self.scale_ratio = 1
-
-    def forward(self, LR):
-#         LR_Feat = self.bn(LR)
-        LR_feat = F.leaky_relu(self.conv1(LR),negative_slope=0.2)
-        LR_feat = F.leaky_relu(self.conv2(LR_feat),negative_slope=0.2)
-        
-        ##Creating Skip connection between dense blocks 
-        out = self.RDB1(LR_feat) 
-#         out = out + LR_feat
-        out1= self.RDB2(out)
-#         out1= out + out1
-        
-        out2 = self.RDB3(out1)
-# #         out2 = out + out2
-        
-#         out2= out2 + LR_feat
-        out3 = self.RDB4(out2)
-        out3= out3 + LR_feat
-#         out3 = out + out1 + out3
-        out4 = self.RDB5(out3)
-        out5 = self.RDB6(out4)
-        out6=torch.cat((out,out1,out2,out3,out4,out5), dim=1)
-        out6=self.conv5(out6)
-        out6= out6.mul(self.scale_ratio) + LR_feat
-        out6 = self.up(out6)
-        #LR_feat = self.resnet(out3)
-        SR=F.leaky_relu(self.conv3(out6),negative_slope=0.2)
-        SR =self.conv4(SR)
-        # print(SR.shape)
-        return SR   
-    
-    
-class ResidualDenseBlock(nn.Module):
-    r"""The residual block structure of traditional SRGAN and Dense model is defined"""
-
-    def __init__(self, channels: int = 64, growth_channels: int = 48, scale_ratio: float = 0.2):
-        """
-
-        Args:
-            channels (int): Number of channels in the input image. (Default: 64).
-            growth_channels (int): how many filters to add each layer (`k` in paper). (Default: 32).
-            scale_ratio (float): Residual channel scaling column. (Default: 0.2)
-        """
-        super(ResidualDenseBlock, self).__init__()
-        self.conv1 = nn.Sequential(
-            nn.Conv2d(channels + 0 * growth_channels, growth_channels, kernel_size=3, stride=1, padding=1),
-            nn.LeakyReLU(negative_slope=0.2, inplace=True)
-#             nn.PReLU()
-        )
-        self.conv2 = nn.Sequential(
-            nn.Conv2d(channels + 1 * growth_channels, growth_channels, kernel_size=3, stride=1, padding=1),
-            nn.LeakyReLU(negative_slope=0.2, inplace=True)
-#             nn.PReLU()
-        )
-        self.conv3 = nn.Sequential(
-            nn.Conv2d(channels + 2 * growth_channels, growth_channels, kernel_size=3, stride=1, padding=1),
-            nn.LeakyReLU(negative_slope=0.2, inplace=True)
-#             nn.PReLU()
-        )
-        self.conv4 = nn.Sequential(
-            nn.Conv2d(channels + 3 * growth_channels, growth_channels, kernel_size=3, stride=1, padding=1),
-            nn.LeakyReLU(negative_slope=0.2, inplace=True)
-#             nn.PReLU()
-        )
-        self.conv5 = nn.Conv2d(channels + 4 * growth_channels, channels, kernel_size=3, stride=1, padding=1)
-
-        self.scale_ratio = scale_ratio
-
-
-    def forward(self, input: torch.Tensor) -> torch.Tensor:
-        conv1 = self.conv1(input)
-        conv2 = self.conv2(torch.cat((input, conv1), 1))
-        conv3 = self.conv3(torch.cat((input, conv1, conv2), 1))
-        conv4 = self.conv4(torch.cat((input, conv1, conv2, conv3), 1))
-        conv5 = self.conv5(torch.cat((input, conv1, conv2, conv3, conv4), 1))
-
-        return conv5.mul(self.scale_ratio) + input
-    
-class ResidualInResidualDenseBlock(nn.Module):
-    r"""The residual block structure of traditional ESRGAN and Dense model is defined"""
-
-    def __init__(self, channels: int = 64, growth_channels: int = 32, scale_ratio: float = 0.2):
-        """
-
-        Args:
-            channels (int): Number of channels in the input image. (Default: 64).
-            growth_channels (int): how many filters to add each layer (`k` in paper). (Default: 32).
-            scale_ratio (float): Residual channel scaling column. (Default: 0.2)
-        """
-        super(ResidualInResidualDenseBlock, self).__init__()
-        self.RDB1 = ResidualDenseBlock(channels, growth_channels, scale_ratio)
-        self.RDB2 = ResidualDenseBlock(channels, growth_channels, scale_ratio)
-        self.RDB3 = ResidualDenseBlock(channels, growth_channels, scale_ratio)
-
-        self.scale_ratio = scale_ratio
-
-    def forward(self, input: torch.Tensor) -> torch.Tensor:
-        out = self.RDB1(input)
-        out = self.RDB2(out)
-        out = self.RDB3(out)
-
-        return out.mul(self.scale_ratio) + input
-    
-
     
 
 def train_discriminator(optimizer, real_data, fake_data,discriminator,b_loss):
@@ -724,37 +341,17 @@ def train_generator(model, optimizer, fake_data,real_data,discriminator,b_loss,m
     loss_perceptual_low=m_loss(features_gt_low, features_out_low)
     loss_perceptual_high=m_loss(features_gt_high, features_out_high)
     
-    ##Image Gradient Loss
-    img_grad = m_loss(calculate_image_gradient(fake_data, real_data))
+#     ##Image Gradient Loss
+#     sobel = SobelGrad()
+#     pred_gradx,pred_grady,label_gradx,label_grady = sobel(fake_data, real_data)
+#     pred_g = torch.sqrt(torch.pow(pred_gradx,2) + torch.pow(pred_gradx,2))
+#     label_g =torch.sqrt(torch.pow(label_gradx,2) + torch.pow(label_grady,2))
+#     img_grad = m_loss(pred_g,label_g)
 
-    total_loss = loss + lambda_*error + 0*loss_perceptual_low + 0*loss_perceptual_high + img_grad
+    total_loss = loss + lambda_*error + 0*loss_perceptual_low + 0*loss_perceptual_high
     total_loss.mean().backward()
     optimizer.step()
-    return loss,error,total_loss,img_grad
-
-def calculate_image_gradient(x,y):
-    sobel_x = torch.Tensor([[1, 0, -1],
-    [2, 0, -2],
-    [1, 0, -1]])
-    
-    sobel_y = torch.Tensor([[1, 2, 1],
-    [0, 0, 0],
-    [-1, -2, -1]])
-    sobel_x = sobel_x.view((1,1,3,3))
-    sobel_y = sobel_y.view((1,1,3,3))
-  
-    #on network output
-    G_x = F.conv2d(x, sobel_x)   
-    G_y = F.conv2d(x, sobel_y)
-    x = torch.sqrt(torch.pow(G_x,2)+ torch.pow(G_y,2))
-    
-    #on ground truth
-    G_x = F.conv2d(y, sobel_x)   
-    G_y = F.conv2d(y, sobel_y)
-    y = torch.sqrt(torch.pow(G_x,2)+ torch.pow(G_y,2))
-    
-    return x,y
-    
+    return loss,error,total_loss,loss_perceptual_low
 
 
 def train_network(trainloader, testloader_flickr,testloader_div,testloader_urban, debug,num_epochs=200,K=10):
